+++ minecraft/net/minecraft/src/mod_statue.java
@@ -1,25 +1,30 @@
 package net.minecraft.src;
 
 import java.util.Random;
+import java.util.List;
 import net.minecraft.client.Minecraft;
 
 public class mod_statue extends BaseMod
 {
-	public static final Block statue = new Statue(128, net.minecraft.src.TileEntity_statue.class).setHardness(1F).setResistance(1F).setBlockName("statue");
+	public static final Block statue = new Statue(128, net.minecraft.src.TileEntity_statue.class, Material.rock).setHardness(1F).setResistance(1F).setBlockName("statue");
 	public static final Item itemStatue = new ItemStatue(500, statue).setIconIndex(ModLoader.addOverride("/gui/items.png", "/dolfinsbizou/itemstatue.png")).setItemName("itemStatue");
 	public static final Item burin = new ItemBurin(501).setIconIndex(ModLoader.addOverride("/gui/items.png", "/dolfinsbizou/burin.png")).setItemName("burin");
 	public static final Item marteau = new ItemMarteau(502).setIconIndex(ModLoader.addOverride("/gui/items.png", "/dolfinsbizou/marteau.png")).setItemName("marteau");
 	public static final Item poudreMagique = new ItemMagicPowder(503).setIconIndex(ModLoader.addOverride("/gui/items.png", "/dolfinsbizou/poudremagique.png")).setItemName("poudreMagique");
+	public static int slotStandardBlock;
 
 	
 	
 	public String getVersion()
 	{
-		return "2r2";
+		return "2r3 skeleton update";
 	}
 	
 	public void load()
 	{
+		ModLoader.setInGUIHook(this, true, true);
+		ModLoader.setInGameHook(this, true, true);
+		 
 		Render_statue render_statue = new Render_statue();
 
 		ModLoader.registerTileEntity(net.minecraft.src.TileEntity_statue.class, "TileEntity_statue", render_statue);
@@ -30,8 +35,8 @@ public class mod_statue extends BaseMod
 		                {
 		            "SSS",
 		            "SPS",
-		            "WWW",
-		Character.valueOf('S'), Block.stone, Character.valueOf('W'), Block.wood, Character.valueOf('P'), mod_statue.poudreMagique
+		            "SSS",
+		Character.valueOf('S'), Block.stone, Character.valueOf('P'), mod_statue.poudreMagique
 		                });
 		
 		ModLoader.addName(burin, "Chisel");
@@ -62,6 +67,39 @@ public class mod_statue extends BaseMod
 		            Block.sand, Item.diamond, Block.wood, 
 		            Item.coal, Item.blazePowder, Block.dirt,
 		                });
+		slotStandardBlock = ModLoader.addOverride("/gui/items.png", "/dolfinsbizou/slotstandardblock.png");
 	}
+	
+	@Override
+    public boolean onTickInGame(float f, Minecraft minecraft)
+    {
+        if(minecraft.currentScreen == null)
+        {
+            lastGuiOpen = null;
+        }
+        return true;
+    }
+ 
+ 
+ 
+   @Override
+    public boolean onTickInGUI(float f, Minecraft minecraft, GuiScreen guiscreen)
+    {
+        if((guiscreen instanceof GuiContainerCreative) && !(lastGuiOpen instanceof GuiContainerCreative))
+        {
+            lastGuiOpen = guiscreen;
+            Container container = ((GuiContainer)guiscreen).inventorySlots;
+            List list = ((ContainerCreative)container).itemList;
+            {
+                list.add(new ItemStack(itemStatue, 1, 0));
+                list.add(new ItemStack(marteau, 1, 0));
+                list.add(new ItemStack(burin, 1, 0));
+                list.add(new ItemStack(poudreMagique, 1, 0));
+            }
+        }
+        return true;
+    }
+ 
+ private static GuiScreen lastGuiOpen;
 	
 }